+++ src/Bicep.Core/Semantics/Metadata/ResourceMetadataCache.cs
@@ -3,63 +3,10 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Immutable;
-using Bicep.Core.Resources;
 using Bicep.Core.Syntax;
-using Bicep.Core.TypeSystem;
 
 namespace Bicep.Core.Semantics.Metadata
 {
-    public class ResourceMetadataParent
-    {
-        public ResourceMetadataParent(ResourceMetadata metadata, SyntaxBase? indexExpression, bool isNested)
-        {
-            // TODO: turn this into a record when the target framework supports it
-            Metadata = metadata;
-            IndexExpression = indexExpression;
-            IsNested = isNested;
-        }
-
-        public ResourceMetadata Metadata { get; }
-
-        public SyntaxBase? IndexExpression { get; }
-
-        public bool IsNested { get; }
-    }
-
-    public class ResourceMetadata
-    {
-        public ResourceMetadata(
-            ResourceType type,
-            SyntaxBase nameSyntax,
-            ResourceSymbol symbol,
-            ResourceMetadataParent? parent,
-            SyntaxBase? scopeSyntax,
-            bool isExistingResource)
-        {
-            // TODO: turn this into a record when the target framework supports it
-            Type = type;
-            NameSyntax = nameSyntax;
-            Symbol = symbol;
-            Parent = parent;
-            ScopeSyntax = scopeSyntax;
-            IsExistingResource = isExistingResource;
-        }
-
-        public ResourceSymbol Symbol { get; }
-
-        public ResourceTypeReference TypeReference => Type.TypeReference;
-
-        public ResourceType Type { get; }
-
-        public ResourceMetadataParent? Parent { get; }
-
-        public SyntaxBase NameSyntax { get; }
-
-        public SyntaxBase? ScopeSyntax { get; }
-
-        public bool IsExistingResource { get; }
-    }
-
     public class ResourceMetadataCache : SyntaxMetadataCacheBase<ResourceMetadata?>
     {
         private readonly SemanticModel semanticModel;
